backend/requirements.txt
flask==3.0.0
tensorflow-cpu>=2.16.1
numpy==1.26.0
werkzeug==3.0.1
flask-cors
gunicorn